LayoutTests/imported/w3c/web-platform-tests/wasm/jsapi/instance/constructor-caching.any-expected.txt


FAIL constructor-caching assert_equals: expected object "[object WebAssembly.Global]" but got object "[object WebAssembly.Global]"
